backend/requirements.txt
asgiref==3.11.0
Django==6.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.3.1
pillow==12.0.0
psycopg[binary]==3.2.13
sqlparse==0.5.5
PyJWT==2.10.1
openpyxl==3.1.5
pytest==8.3.2
pytest-django==4.8.0
